Be khoa/feat/patient management

+++ be/internal/domain/entity/medicine.go
@@ -3,10 +3,10 @@ package entity
 import "time"
 
 type Medicine struct {
-	Id              int64  `gorm:"primaryKey;autoIncrement"`
-	Name            string `gorm:"type:varchar(256)"`
-	UsesInstruction string `gorm:"type:varchar(256)"`
-	Price           int    `json:"price"`
+	Id              int64   `gorm:"primaryKey;autoIncrement"`
+	Name            string  `gorm:"type:varchar(256)"`
+	UsesInstruction string  `gorm:"type:varchar(256)"`
+	Price           float32 `json:"price"`
 	CreatedAt       time.Time
 	UpdatedAt       time.Time
 }
